+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## Changelog
 
+### 2016.07.22
+* `form_options`, `form_attributes` and `form_url` methods moved from Releaf::ActionController to `Releaf::Builders::EditBuilder` as its edit builder responsibility for it's own content
+* `Releaf::Builders::EditBuilder` for now have `#form_builder_class`
+  method for custom form builder class overriding
+* `table_options` method moved from Releaf::ActionController to `Releaf::Builders::IndexBuilder`
+
 ### 2016.07.19
 * `Releaf::ItemOrderer` refactored to `Array::Reorder` service
 * Array now have #reorder method for simple array reordering